refactor: User 도메인 리팩토링 - points -> point
