+++ src/main/java/com/flab/just_10_minutes/User/dto/UserDto.java
@@ -32,7 +32,7 @@ public class UserDto{
                 .phone(userDto.getPhone())
                 .address(userDto.getAddress())
                 .role(userDto.getRole())
-                .points(0L)
+                .point(0L)
                 .build();
     }
 }
