+++ src/test/java/com/flab/just_10_minutes/Point/fixture/PointHistoryDtoTestFixture.java
