+++ src/test/java/com/flab/just_10_minutes/Point/fixture/PointHistoryTestFixture.java
