+++ src/test/java/com/flab/just_10_minutes/Point/infrastructure/PointDaoTest.java
