+++ src/test/java/com/flab/just_10_minutes/Point/service/PointServiceTest.java
